Complete ModelSpec #issue46
